Fix little details
Instead of a no_soln variable, we can just return.

Break out of loop earlier if curr_node is target, it makes more sense
not to visit its neighbours.

No need to special case start=end, the backtracking code works for
that case perfectly.

Uncomment the sanity check, useful.

Fix assert to check both functions if No solution exists, previously
it was not taken in account.